modification du code, simplification et ajout de la branch dev
