+++ src/composants/ButtonFavorites/ButtonFavorites.css
@@ -1,0 +1,9 @@
+.heart-red {
+    background-color: none;
+    color: rgb(255, 0, 0);
+  }
+  
+  .heart-bg {
+    background-color: none;
+    color: rgb(0, 0, 0);
+  }
