+++ src/composants/CategorieBar/categorie.css
@@ -19,7 +19,7 @@
     padding-top: 1rem;
     height: 5.125rem;
     position: fixed;
-    width: 100%;
+    max-width: 100%;
     top: 3.9rem;
   }
