+++ src/composants/HyruleMap/hyrulemap.css
@@ -34,7 +34,7 @@ section{
     min-width: 70vw;
     border-radius: 10px;
     width: 100%;
-       max-width: 1920px;   
+    max-width: 1920px;   
      min-height: 50vh;
      margin-right: auto;
      margin-left: auto;
