+++ src/index.css
@@ -57,7 +57,9 @@ a:hover {
 
 
 body {
+  background-color: var(--main-color);
   margin: 0;
+  display: flex;
   place-items: center;
   min-width: 320px;
   min-height: 100vh;
